requirements.txt
adjustText==0.7.3
asgiref==3.5.0
blank==0.1.0
certifi==2021.10.8
comments==1.0.0
cycler==0.11.0
decorator==5.1.1
Django==4.0.4
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-js-asset==2.0.0
django-mptt==0.13.4
django-stubs==1.10.1
django-stubs-ext==0.4.0
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
fonttools==4.33.2
future==0.18.2
iso8601==1.0.2
joblib==1.1.0
kiwisolver==1.4.2
many==0.6.9
matplotlib==3.5.1
mypy==0.942
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pandas==1.4.2
patsy==0.5.2
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
self==2020.12.3
serial==0.0.97
six==1.16.0
sklearn==0.0
slug==2.0
sqlparse==0.4.2
statsmodels==0.13.2
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.0
types-pytz==2021.3.6
types-PyYAML==6.0.7
typing_extensions==4.2.0
urllib3==1.26.9
